requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
binaryornot==0.4.4
black==24.2.0
certifi==2023.11.17
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
distlib==0.3.8
Django==4.2.3
django-adminlte-3==0.1.6
django-auto-prefetch==1.8.0
django-auto-prefetching==0.2.12
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-resized==1.0.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==23.2.0
filelock==3.13.1
flake8==7.0.0
identify==2.5.33
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
pre-commit==3.6.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pytest==8.0.0
pytest-django==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
setuptools==69.0.3
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
